bug: multi-record Key/Key URL favicon display error
When a Key/Key URL key/value set is being used, and there can be multiple entities in this key set (i.e. the original data was split), favicon icons are being assigned to the last Key record (i.e. the recursively found N-1 record). 

Because the code progresses through the key and key URLs in order, and the order for multi-entries are Key (1), Key (2), Key (3), Key URL (1), Key URL (2), Key URL (3), the following logic was used. Search for all keys with the correct text (logic unchanged), AND where the value TD (td:nth-child(2)) does not have a "span.favicon" element, as adding the span.favicon element indicates that a Key/Key URL has been processed. Finally, return only the first result as the $target.